ASHSH-1015: Add gpio input to detect a signal
When the garage door was activated there was no way for outside
developers to confirm that the garage door opener actually received a
signal. In order to notify the caller that the garage door opener
recieved a signal an input pin was added to receive some of the signal
and notify the caller that the garage door opener had really been
activated.

Added a config file to store the pin configuration based on environment.
Only the development environment is setup right now.

Refactored some of the code to use the updated python "with" function.
This made it so that an object with an __enter__() and __exit__()
function can do the setup and cleanup much easier than a "try" "finally"
block.

Refactored some of the wording as well.

Also the relay was removed from the circuit and a simple NPN transistor
was added. This made it much easier to control the garage circuit.